Clean and organize the code
