--- src/main/java/com/wilsonfranca/saintseya/SaintSeyaGame.java
@@ -1,12 +1,0 @@
-package com.wilsonfranca.saintseya;
-
-/**
- * Created by wilson.franca on 06/04/18.
- */
-public class SaintSeyaGame {
-
-    public static void main(String args[]) {
-        Game game = new Game();
-        game.play();
-    }
-}
