+++ src/main/java/com/wilsonfranca/saintseya/battle/Reward.java
@@ -1,4 +1,4 @@
-package com.wilsonfranca.saintseya;
+package com.wilsonfranca.saintseya.battle;
 
 import java.util.Arrays;
