+++ src/main/java/com/wilsonfranca/saintseya/campaign/CampaignController.java
@@ -1,8 +1,7 @@
 package com.wilsonfranca.saintseya.campaign;
 
 import com.wilsonfranca.saintseya.GameEngine;
-import com.wilsonfranca.saintseya.Player;
-import com.wilsonfranca.saintseya.Quest;
+import com.wilsonfranca.saintseya.player.Player;
 
 /**
  * Created by wilson on 15/04/18.
@@ -19,7 +18,4 @@ public class CampaignController {
         gameEngine.createKnight(player);
     }
 
-    public void execute(Quest quest) {
-        gameEngine.startQuest(quest);
-    }
 }
